% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method_checkReferenceAnchor.R
\name{.checkReferenceAnchor}
\alias{.checkReferenceAnchor}
\title{Checks if mutations and reference genome correspond}
\usage{
.checkReferenceAnchor(
  mutations,
  genomeSeqs,
  ignoreNonMatch = FALSE,
  threads = 1
)
}
\arguments{
\item{mutations}{(\link[VariantAnnotation]{VRanges}): Mutations to incorporate within transcripts.}

\item{genomeSeqs}{(\link[BSgenome]{BSgenome} or \link[Biostrings]{DNAStringSet}): Reference genome sequences.}

\item{ignoreNonMatch}{(logical): Should non-matching reference anchors be ignored? These mutations will be removed.}

\item{threads}{(integer): Number of threads.}
}
\value{
(logical): Returns TRUE/FALSE if per given mutation based on matching position and base with the reference genome.
}
\description{
Overlaps the reference bases and positions with the given genome and check if all positions overlap and reports
non-overlapping mutations.
}
\examples{

ProteoDiscography.hg19 <- ProteoDisco::generateProteoDiscography(
  TxDb = TxDb.Hsapiens.UCSC.hg19.knownGene::TxDb.Hsapiens.UCSC.hg19.knownGene, 
  genomeSeqs = BSgenome.Hsapiens.UCSC.hg19::BSgenome.Hsapiens.UCSC.hg19
)

# Supply the ProteoDiscography with genomic variants to incorporate in downstream analysis. This can be one or multiple VCF / MAF files.
# Additional manual sequences and exon-exon mapping (i.e., splice junctions) can also be given as shown in the sections below.
ProteoDiscography.hg19 <- ProteoDisco::importGenomicVariants(
  ProteoDiscography = ProteoDiscography.hg19,
  # Provide the VCF / MAF files, if more then one supply a vector of files and corresponding samplenames.
  files = system.file('extdata', 'validationSet_hg19.vcf', package = 'ProteoDisco'), 
  # We can replace the original samples within the VCF with nicer names.
  samplenames = 'Validation Set (GRCh37)',
  # Number of threads used for parallelization.
  # We run samples sequentially and parallelize within (variant-wise multi-threading).
  threads = 1, 
  # To increase import-speed for this example, do not check for validity of the reference anchor with the given reference sequences.
  performAnchorCheck = FALSE
)

}
\author{
Job van Riet \email{j.vanriet@erasmusmc.nl}

Wesley van de Geer \email{w.vandegeer@erasmusmc.nl}
}
\concept{methods}
\keyword{methods}
