% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/AllMethods.R
\name{print}
\alias{print}
\alias{print,ProteoDiscography-method}
\title{Display a summary of ProteoDiscography}
\usage{
print(object)

\S4method{print}{ProteoDiscography}(object)
}
\arguments{
\item{object}{(ProteoDiscography): ProteoDiscography object.}
}
\value{
Displays information relating to the ProteoDiscography into the console.
}
\description{
Display a summary of ProteoDiscography
}
\examples{

# Import example ProteoDiscography (hg19) and re-link TxDb.
data('ProteoDiscographyExample.hg19', package = 'ProteoDisco')
ProteoDiscographyExample.hg19@TxDb <- TxDb.Hsapiens.UCSC.hg19.knownGene::TxDb.Hsapiens.UCSC.hg19.knownGene

print(ProteoDiscographyExample.hg19)

}
