% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/AllMethods.R
\name{seqinfo}
\alias{seqinfo}
\alias{seqinfo,ProteoDiscography-method}
\title{Retrieve seqinfo.}
\usage{
seqinfo(x)

\S4method{seqinfo}{ProteoDiscography}(x)
}
\arguments{
\item{x}{(ProteoDiscography): ProteoDiscography object.}
}
\value{
Returns an Seqinfo-class containing information on the imported genomic sequences.
}
\description{
Retrieve seqinfo.
}
\examples{

# Import example ProteoDiscography (hg19)
data('ProteoDiscographyExample.hg19', package = 'ProteoDisco')
ProteoDiscographyExample.hg19 <- setTxDb(ProteoDiscographyExample.hg19, TxDb.Hsapiens.UCSC.hg19.knownGene::TxDb.Hsapiens.UCSC.hg19.knownGene)
ProteoDiscographyExample.hg19 <- setGenomicSequences(ProteoDiscographyExample.hg19, BSgenome.Hsapiens.UCSC.hg19::BSgenome.Hsapiens.UCSC.hg19)

seqinfo(ProteoDiscographyExample.hg19)

}
