% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/AllMethods.R
\name{seqlevels}
\alias{seqlevels}
\alias{seqlevels,ProteoDiscography-method}
\title{Retrieve seqlevels.}
\usage{
seqlevels(x)

\S4method{seqlevels}{ProteoDiscography}(x)
}
\arguments{
\item{x}{(ProteoDiscography): ProteoDiscography object.}
}
\value{
seqlevels of the improted genomic sequences.
}
\description{
Retrieve seqlevels.
}
\examples{

# Import example ProteoDiscography (hg19) and re-link TxDb.
data('ProteoDiscographyExample.hg19', package = 'ProteoDisco')
ProteoDiscographyExample.hg19@TxDb <- TxDb.Hsapiens.UCSC.hg19.knownGene::TxDb.Hsapiens.UCSC.hg19.knownGene

seqlevels(ProteoDiscographyExample.hg19)

}
