% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/AllMethods.R
\name{show}
\alias{show}
\alias{show,ProteoDiscography-method}
\title{Display a summary of ProteoDiscography}
\usage{
show(object)

\S4method{show}{ProteoDiscography}(object)
}
\arguments{
\item{object}{(ProteoDiscography): ProteoDiscography object.}
}
\value{
Displays information relating to the ProteoDiscography into the console.
}
\description{
Display a summary of ProteoDiscography
}
\examples{

ProteoDiscography.hg19 <- ProteoDisco::generateProteoDiscography(
  TxDb = TxDb.Hsapiens.UCSC.hg19.knownGene::TxDb.Hsapiens.UCSC.hg19.knownGene,
  genomeSeqs = BSgenome.Hsapiens.UCSC.hg19::BSgenome.Hsapiens.UCSC.hg19
)

show(ProteoDiscography.hg19)

}
