% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/AllMethods.R
\name{organism}
\alias{organism}
\alias{organism,ProteoDiscography-method}
\title{Display the organism of the TxDb}
\usage{
organism(object)

\S4method{organism}{ProteoDiscography}(object)
}
\arguments{
\item{object}{(ProteoDiscography): ProteoDiscography object.}
}
\value{
Returns the specified organism during creation.
}
\description{
Display the organism of the TxDb
}
\examples{

# Import example ProteoDiscography (hg19) and re-link TxDb.
data('ProteoDiscographyExample.hg19', package = 'ProteoDisco')
ProteoDiscographyExample.hg19@TxDb <- TxDb.Hsapiens.UCSC.hg19.knownGene::TxDb.Hsapiens.UCSC.hg19.knownGene

organism(ProteoDiscographyExample.hg19)

}
