% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/AllMethods.R
\name{print}
\alias{print}
\alias{print,ProteoDiscography-method}
\title{Display a summary of ProteoDiscography}
\usage{
print(object)

\S4method{print}{ProteoDiscography}(object)
}
\arguments{
\item{object}{(ProteoDiscography): ProteoDiscography object.}
}
\value{
Displays information relating to the ProteoDiscography into the console.
}
\description{
Display a summary of ProteoDiscography
}
\examples{

# Import example ProteoDiscography (hg19)
data('ProteoDiscographyExample.hg19', package = 'ProteoDisco')
ProteoDiscographyExample.hg19 <- setTxDb(ProteoDiscographyExample.hg19, TxDb.Hsapiens.UCSC.hg19.knownGene::TxDb.Hsapiens.UCSC.hg19.knownGene)
ProteoDiscographyExample.hg19 <- setGenomicSequences(ProteoDiscographyExample.hg19, BSgenome.Hsapiens.UCSC.hg19::BSgenome.Hsapiens.UCSC.hg19)

print(ProteoDiscographyExample.hg19)

}
