% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/AllMethods.R
\name{seqinfo}
\alias{seqinfo}
\alias{seqinfo,ProteoDiscography-method}
\title{Retrieve seqinfo.}
\usage{
seqinfo(x)

\S4method{seqinfo}{ProteoDiscography}(x)
}
\arguments{
\item{x}{(ProteoDiscography): ProteoDiscography object.}
}
\value{
Returns an Seqinfo-class containing information on the imported genomic sequences.
}
\description{
Retrieve seqinfo.
}
\examples{

ProteoDiscography.hg19 <- ProteoDisco::generateProteoDiscography(
  TxDb = TxDb.Hsapiens.UCSC.hg19.knownGene::TxDb.Hsapiens.UCSC.hg19.knownGene,
  genomeSeqs = BSgenome.Hsapiens.UCSC.hg19::BSgenome.Hsapiens.UCSC.hg19
)

seqinfo(ProteoDiscography.hg19)

}
