% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/AllMethods.R
\name{seqlevels}
\alias{seqlevels}
\alias{seqlevels,ProteoDiscography-method}
\title{Retrieve seqlevels.}
\usage{
seqlevels(x)

\S4method{seqlevels}{ProteoDiscography}(x)
}
\arguments{
\item{x}{(ProteoDiscography): ProteoDiscography object.}
}
\value{
seqlevels of the improted genomic sequences.
}
\description{
Retrieve seqlevels.
}
\examples{

ProteoDiscography.hg19 <- ProteoDisco::generateProteoDiscography(
  TxDb = TxDb.Hsapiens.UCSC.hg19.knownGene::TxDb.Hsapiens.UCSC.hg19.knownGene,
  genomeSeqs = BSgenome.Hsapiens.UCSC.hg19::BSgenome.Hsapiens.UCSC.hg19
)

seqlevels(ProteoDiscography.hg19)

}
