% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/AllMethods.R
\name{summary}
\alias{summary}
\alias{summary,ProteoDiscography-method}
\title{Display a summary of ProteoDiscography}
\usage{
summary(object, verbose = TRUE)

\S4method{summary}{ProteoDiscography}(object, verbose = TRUE)
}
\arguments{
\item{object}{(ProteoDiscography): ProteoDiscography object.}

\item{verbose}{(logical): Set the verbosity}
}
\value{
Displays information relating to the ProteoDiscography into the console.
If verbose is set to TRUE, it will also output the imported genomic variants, splicing junctions and manual sequences as a list of tibbles.
}
\description{
Display a summary of ProteoDiscography
}
\examples{

# Import example ProteoDiscography (hg19)
data('ProteoDiscographyExample.hg19', package = 'ProteoDisco')
ProteoDiscographyExample.hg19 <- setTxDb(ProteoDiscographyExample.hg19, TxDb.Hsapiens.UCSC.hg19.knownGene::TxDb.Hsapiens.UCSC.hg19.knownGene)
ProteoDiscographyExample.hg19 <- setGenomicSequences(ProteoDiscographyExample.hg19, BSgenome.Hsapiens.UCSC.hg19::BSgenome.Hsapiens.UCSC.hg19)

summary(ProteoDiscographyExample.hg19)

}
