% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/AllMethods.R
\name{summary}
\alias{summary}
\alias{summary,ProteoDiscography-method}
\title{Display a summary of ProteoDiscography}
\usage{
summary(object, verbose = TRUE)

\S4method{summary}{ProteoDiscography}(object, verbose = TRUE)
}
\arguments{
\item{object}{(ProteoDiscography): ProteoDiscography object.}

\item{verbose}{(logical): Set the verbosity}
}
\value{
Displays information relating to the ProteoDiscography into the console.
If verbose is set to TRUE, it will also output the imported genomic variants, splicing junctions and manual sequences as a list of tibbles.
}
\description{
Display a summary of ProteoDiscography
}
\examples{

ProteoDiscography.hg19 <- ProteoDisco::generateProteoDiscography(
  TxDb = TxDb.Hsapiens.UCSC.hg19.knownGene::TxDb.Hsapiens.UCSC.hg19.knownGene, 
  genomeSeqs = BSgenome.Hsapiens.UCSC.hg19::BSgenome.Hsapiens.UCSC.hg19
)

summary(ProteoDiscography.hg19)

}
